Automatically send a message on Community Day
Adds an extra bit to the bg_task which first gets a proper Channel
object from a given ID. You'll need to replace that by the way.

After that, we create a "while loop" which runs forever, as long as
the PgsBot (self) connection is open (*not* is_closed()).

Grab the current time/date and call it "now", and set up somewhere to
store the next date of community day too. Call it nextCommunityDay.
If the next event is 15th August, you'd set it to 15.

Finally, check if the current day matches the next Community Day.
If it does, we can send a message, but we need to make sure it's the
right time too, so we also check hour and minute is 8/0 to run at 08:00.

To stop the bot from sending a flood of messages for a whole minute, we
pause at the end for 60 seconds to make sure only one message gets sent.