Send a message when users send a message
Sends the literal text "Thanks for your message!" by subscribing to the
on_message event which is triggered by the Discord API every time
somebody sends a message.

The message object we get back from Discord in that event will contain
the channel that the message was sent from. We can use that to make sure
the message gets sent to the right place.