Talking to yourself is the first sign of madness.
Yup, the bot responds to its own messages too. Luckily, there's a way to
stop the infinite spam once anyone posts a message to the channel.

By comparing the Author ID from the Message object to the User ID that
the bot is logged in as (remember, self always refers to the bot class)
we can bail out if the message was posted by the bot.

Once you `return`, nothing else in that method will run, so when it sees
it's a message sent by itself, it essentially sees it and does nothing.