Add our first command
Gets rid of the thankyou spam and replaces it with something more useful

Check each sent message to see if it matches "!ping".
If it does, we say "pong". Simple as that!

Seems like a whole lot of nothing, but this can actually be useful if
you ever want to see if the bot is broken or Discord's being a dick.