Add loading/connected status to console output
As soon as the PgsBot class is initialised, we output a loading message
to the console window so we know it's doing something.

We then create a background task which will run immediately after. Once
the bg_task method sees that Discord is ready to rock, it now tells us.