Add timed, alternating community day message
First check the weekday - which starts at 0 being Monday, 1 for Tuesday
and so on. So, when it's Thursday we want to post a message at 10:00.

Nest migrations only happen every *other* week though, so to alternate
the message each week we can grab the week of the year with the
`isocalendar()[1]` part, then use `% 2` to get the remainder after
diving the week number by 2. The result will only ever be 1 or 0.