Add helper method for getting channels
To get a channel we need an integer which, because the ConfigParser sees
everything as strings, needs extra code.

Rather than running `self.config['channels'].getint('some-channel')`
every time, this adds a helper so we can run `self.get_channel('foo')`

If we pass something to get_channel that's not an integer (int), we
grab that channel from the config and convert the textual ID to an int.
After that, we defer back to the "parent" discord.Client class with the
`super().get_channel()` bit. That's because we're overriding the normal
get_channel method, so if we don't then we won't get anything.