Add a "default" channel in place of hard-coded one
Rather than having "testing" locked in in the code (it mostly won't be
the testing channel, it'll be a real channel when the bot goes live),
we'll call it "default" and use it as a general fallback.

The get_channel method is updated to check if no channel was passed at
all, and sets it to "default" if that's the case. So, for example:

self.get_channel('nests') // Get the 'nests' channel from config
self.get_channel() // Same thing as self.get_channel('default')