Add more granular control of community day config
Splits the response text of the !community command so that instead of
specifying "**Squirtle** on Sunday **8th June**" the day, date and
pokemon are all set individually.

Not only does it mean you don't need to piss about with Markdown in the
config file, it gives more flexibility over what we can do with other
stuff too. We can't work with the date if it starts with Squirtle!