Add !migrating command for prev/next nest change
First, we need to reuse the isocalendar bit that checks whether we're on
an odd or even week, so we've moved it to an `is_migration_week` method.
Now we can use self.is_migration_week() instead of copying code everywhere

The next_migration method first checks to see if it isn't even the week
of a migration. If it IS migration week, then we need to check if it's
before or after Thursday - so that's what the first two `elif` lines do.

After that point, we know it's Thursday and it's a migration week, so
it's just a matter of pumping out different text based on current time.