Add test_mode option to divert messages
With test mode on, all timed messages will be diverted to the channel
set as 'testing' in the channels list.

With test mode off, timed messages will go to the respective channels
such as #event-discussions for the community day timed messages.

Command responses are not affected, they will always respond to the
channel the command was invoked from.