Set online and playing status for the bot
Pretty much does what it says on the tin. change_presence() is a method
on the `discord.Client` class we extend, so we can get to it through
`self`. We pass two named parameters in, one to set the status online.

Status doesn't really need to be set, online is the default, but it
makes it easier to change if you should ever decide on 'idle'.
All the available status options you can set are at this link:
https://discordpy.readthedocs.io/en/rewrite/api.html#discord.Status

We set activity by passing a custom `discord.Game` data class as the
value for the `activity` option. It only takes one option for bots, and
that's the name, which we make up ourselves.