Add "Repeat after me" command
First thing we do is send a message so you know when the bot's ready.

Once that's sent, the bot will start waiting for another message to be
posted. If that message's author matches the original message author,
then we're good to go.

Last step is to send the content of the reply to the same channel that
the original message was sent from.