Ask which channel the message should be posted to
Basically the same as what's already there, but instead of just checking
the author matches on the reply, it also checks that exactly ONE channel
is mentioned in the reply. If you mention two, it'll just keep waiting.

The only other difference is we're now sending the content of the first
reply to the first channel (starting at 0) mentioned in the 2nd reply.