Automatically delete trigger and q/a messages
Assigns the "What should I say?" and "Where should I post it?" messages
to variables (where and what) so that we can work on them afterwards.

Once the message is sent, we loop through each message that we sent or
received, assigning each one to the `ch` variable. In other words, the
`for ch in [...]:` line followed by `await ch.delete()` is another way
of running the following block of code:

```
await channel_reply.delete()
await where.delete()
await reply.delete()
await what.delete()
await message.delete()
```

We want the success message to stick around for a bit though, so before
we delete that one we tell python to go to sleep for 2 seconds and
concentrate on getting something else done.