Restrict 'repeat after me' to users in admins.txt
You'll need to create an admins.txt file in the project root (in the
same folder as pgsbot.py) and add permitted Member IDs, one per line.

First, we call our new `load_admins` method from `__init__`, i.e. as
soon as the bot is ran in a terminal. That method creates a new list
called `self.admins`, then **tries** to open the admins.txt file.

If the file doesn't exist, we print a message to the console. If it
*does* exist, then `f.read().splitlines()` reads the file we opened,
splitting it up at the end of each line to get a `list` of the lines.

Then it runs `self.add_admin()` **for** every item in the list we got
from `f.read().splitlines()`. The `add_admin()` method will then simply
print out a message to the console so we can know when it's adding
admins, then **appends** the admin to the `self.admins` list, making
sure that it's an integer in the process. If it's not an `int`, Discord
won't like it.

Finally, before running the "repeat after me" command, we to see if the
message author has permission to use it, or `is_admin()`. **If not**,
then we will just `return` which stops the code going any further.