Move !invite and !ping commands to a "misc" cog
A 'cog' is essentially a way to group a bunch of commands and events
(like on_message / on_ready) all together when they're similar.

Think of cogs as part of a chain, it's easier to remember.

This removes the !invite and !ping commands, instead moving them to
cogs/misc.py. In out PgsBot's `__init__` function we tell the parent
commands.Bot class to load the extension called "cogs.misc" - the folder
and the cog's filename (without extension) separated by a dot.

The only other thing to do is @super().on_message() which makes sure ALL
commands get processed, not just the ones in our on_message event, which
is what it would be otherwise.

Long story short, commands will now be easier to write, each organised
neatly as its own method in a dedicated class.

We also add the `__pycache__` folders to the ignore list, as python's
started creating compiled class caches now it seems.