Create the community-day Cog
Same as with the misc, but this is only one command for now.

We now use `ctx.send` instead of `message.channel.send`, making the line
a lot shorter thanks to the send helper in the `ctx` (context) parameter

With shorter lines we can fit more stuff, so we won't bother setting the
text to a `communityText` variable any more either.

When we `__init__`(...ialise) the Cog in the `setup` definition at the
bottom, we receive and pass on the `bot` argument. `bot` points directly
to our `PgsBot` class, so we have access to the config through that.