Move !migrating command to Nests cog
Should be getting the hang of this by now. The `next_migration` def is
moved as-is, but because `self.is_migration_week` would look for the
`is_migration_week` def on `self` - the `Nests` cog - it won't work.

Reason being, `is_migration_week` is still `def`ined in the `PgsBot`
class because we need it for the timed messages. To solve that, we
simply add a line in `Nests.__init__()` which references it for us.

The only difference on the command itself (short of the usual `ctx`
replacements) is that we've added a "migration" alias. With the alias
defined, we'll be able to call it with either !migrating OR !migration.