Move admin helpers and 'repeat after me' command
This one's a little different. Repeat after me isn't *technically* a
command, it's just the bot seeing some text it likes and acting on it.

A true command requires either a '!' prefix, or to @mention the bot. It
can also only be one word, which doesn't work so well for this one.

So, we're basically porting it over as-is, `on_message` event 'n' all.