Add !say command based on 'repeat after me'
Adds a new !say command that takes two required parameters.

The first must be a channel: It can be a mention, an ID or a name, as
long as it actually exists. The second is the message, and the `*`
before it in the `def say(...)` line says "everything after the previous
parameter should be dumped together as one big value in the next param".

Most of the code that handled the "repeat after me" command is now moved
into a new `say_to` definition which takes the context (or `message`),
along with the channel and the text to send. The final argument,
`deletes`, is a list of messages we/they sent that will get deleted.