Make !say command work in DMs
The issue of mentions not being included with DMs is not present in the
!say command as the named parameter in the definition is automatically
converting it to a `discord.Channel`, so we can safely use !say in DMs.

We don't delete any messages though, so we'll return early instead. We
can't delete the other person's messages, so deleting our own would make
things look super weird and might confuse people.