Move weekly nest alert into Nests Cog
is_migration_week was only used by the weekly message in the main pgsbot
file since we already moved the !migration command earlier, so now we
can move it into the Cog with everything else. Less clutter in pgsbot!

Since the main loop is in pgsbot and not our cog, we have to add another
task in the `__init__` method. We'll call it migration_notifier.

It pretty much does the same as before, except to make it easier to read
the `migration_notifier` method we've moved the actual message into its
own separate definition. All that does is literally set the message
based on whether or not it's migration week, which we pass in, then
sends it to whatever the 'nests' channel is set to using get_channel