Fix some minor bugs
discord.status won't work with a little s. Status is a name, it should
be capitalised as such.

Also, when calling the parent's (super) get_channel method, we don't
need to pass self. It goes along automatically.

Not sure how things worked with it this way before, but there you go.